Tidy up unused code and standardise validation responses

+++ app/Http/Controllers/AnswersController.php
@@ -26,11 +26,21 @@ class AnswersController extends Controller
     
     public function create(Question $question, Request $request)
     {
-        $validated = $this->validate($request, [
-            'content' => 'required'
-        ]);
+        try {
+            $this->validate($request, [
+                'content' => 'required'
+            ]);
+        } catch (ValidationException $e) {
+            $errors = $e->validator->errors();
+            
+            return response()->json(['message' => 'Failed to create answer', 'errors' => $errors], 400);
+        }
         
         $loggedInUser = $request->user();
+        
+        if (!$loggedInUser->hasPrivilege('answer:create')) {
+            return response()->json(['message' => 'You do not have permission to answer questions'], 401);
+        }
         
         $question->answers()->create([
             'content' => $request->input('content'),
@@ -43,10 +53,16 @@ class AnswersController extends Controller
     
     public function update(Question $question, Request $request)
     {
-        $validated = $this->validate($request, [
-            'id' => 'required|exists:answers',
-            'content' => 'required'
-        ]);
+        try {
+            $this->validate($request, [
+                'id' => 'required|exists:answers',
+                'content' => 'required'
+            ]);
+        } catch (ValidationException $e) {
+            $errors = $e->validator->errors();
+            
+            return response()->json(['message' => 'Failed to update answer', 'errors' => $errors], 400);
+        }
         
         $loggedInUser = $request->user();
         
@@ -54,6 +70,10 @@ class AnswersController extends Controller
         
         if (empty($answer)) {
             return response()->json(['message' => 'Answer not found'], 404);
+        }
+        
+        if (($answer->created_by === $loggedInUser->id && !$loggedInUser->hasPrivilege('answer:update_self')) && !$loggedInUser->hasPrivilege('answer:update_other')) {
+            return response()->json(['message' => 'You do not have permission to update this answer.'], 401);
         }
         
         $answer->fill($request->all() + ['updated_by' => $loggedInUser->id]);
@@ -65,14 +85,25 @@ class AnswersController extends Controller
     
     public function remove(Question $question, Request $request)
     {
-        $validated = $this->validate($request, [
-            'id' => 'required|exists:answers'
-        ]);
+        try {
+            $this->validate($request, [
+                'id' => 'required|exists:answers'
+            ]);
+        } catch (ValidationException $e) {
+            $errors = $e->validator->errors();
+            
+            return response()->json(['message' => 'Failed to delete answer', 'errors' => $errors], 400);
+        }
         
         $answer = $question->answers()->find($request->input('id'));
-        
+
         if (empty($answer)) {
             return response()->json(['message' => 'Answer not found'], 404);
+        
+        $loggedInUser = $request->user();
+        
+        if (($answer->created_by === $loggedInUser->id && !$loggedInUser->hasPrivilege('answer:delete_self')) && !$loggedInUser->hasPrivilege('answer:delete_other')) {
+            return response()->json(['message' => 'You do not have permission to delete this answer.'], 401);
         }
         
         $answer->delete();
@@ -84,8 +115,8 @@ class AnswersController extends Controller
     {
         $user = $request->user();
         
-        if ($user->id !== $question->id || !$user->hasPrivilege('answer:mark_solution')) {
-            return response()->json(['message' => 'You do not have permission to mark this answer as the solution'], 400);
+        if ($user->id !== $question->created_by || !$user->hasPrivilege('answer:mark_solution')) {
+            return response()->json(['message' => 'You do not have permission to mark this answer as the solution'], 401);
         }
         
         if ($answer->solution === true) {
@@ -103,14 +134,20 @@ class AnswersController extends Controller
     
     public function addVote(Question $question, Answer $answer, Request $request)
     {
-        $this->validate($request, [
-            'direction' => 'required|in:up,down'
-        ]);
+        try {
+            $this->validate($request, [
+                'direction' => 'required|in:up,down'
+            ]);
+        } catch (ValidationException $e) {
+            $errors = $e->validator->errors();
+            
+            return response()->json(['message' => 'Failed to vote on answer', 'errors' => $errors], 400);
+        }
         
         $user = $request->user();
         
         if (!$user->hasPrivilege('answer:vote')) {
-            return response()->json(['message' => 'You do not have permission to vote on answers'], 400);
+            return response()->json(['message' => 'You do not have permission to vote on answers'], 401);
         }
         
         $this->undoExistingVotes($answer, $user);
@@ -136,7 +173,7 @@ class AnswersController extends Controller
         $user = $request->user();
         
         if (!$user->hasPrivilege('answer:vote')) {
-            return response()->json(['message' => 'You do not have permission to vote on answers'], 400);
+            return response()->json(['message' => 'You do not have permission to vote on answers'], 401);
         }
         
         $this->undoExistingVotes($answer, $user);
